feature: Texture implemented (only works for PNG file when reading texture file and float_resize_up() is not working yet)

+++ src/scene/renderer.rs
@@ -52,7 +52,7 @@ fn single_thread_render(
                     }
                 }
 
-                rendered_y.push(y);
+                rendered_y.push(y as usize);
             }
         }
     }
